--- a/src/ReverseProxy/Service/RuntimeModel/Transforms/StructuredTransformer.cs
+++ b/src/ReverseProxy/Service/RuntimeModel/Transforms/StructuredTransformer.cs
@@ -63,12 +63,52 @@ namespace Microsoft.ReverseProxy.Service.RuntimeModel.Transforms
         /// </summary>
         internal IList<ResponseTrailersTransform> ResponseTrailerTransforms { get; }
 
-        public override async Task TransformRequestAsync(HttpContext httpContext, HttpRequestMessage proxyRequest, string destinationPrefix)
+        public override Task TransformRequestAsync(HttpContext httpContext, HttpRequestMessage proxyRequest, string destinationPrefix)
         {
             if (ShouldCopyRequestHeaders.GetValueOrDefault(true))
             {
-                await base.TransformRequestAsync(httpContext, proxyRequest, destinationPrefix);
+                var task = base.TransformRequestAsync(httpContext, proxyRequest, destinationPrefix);
+                if (!task.IsCompletedSuccessfully)
+                {
+                    return AwaitTransformRequestAsync(task, httpContext, proxyRequest, destinationPrefix);
+                }
             }
+
+            if (RequestTransforms.Count == 0)
+            {
+                return Task.CompletedTask;
+            }
+
+            var transformContext = new RequestTransformContext()
+            {
+                DestinationPrefix = destinationPrefix,
+                HttpContext = httpContext,
+                ProxyRequest = proxyRequest,
+                Path = httpContext.Request.Path,
+                Query = new QueryTransformContext(httpContext.Request),
+                HeadersCopied = ShouldCopyRequestHeaders.GetValueOrDefault(true),
+            };
+
+            for (var i = 0; i < RequestTransforms.Count; i++)
+            {
+                var task = RequestTransforms[i].ApplyAsync(transformContext);
+                if (!task.IsCompletedSuccessfully)
+                {
+                    return AwaitTransformRequestAsync(task, i, transformContext, proxyRequest);
+                }
+                task.GetAwaiter().GetResult();
+            }
+
+            // Allow a transform to directly set a custom RequestUri.
+            proxyRequest.RequestUri ??= RequestUtilities.MakeDestinationAddress(
+                transformContext.DestinationPrefix, transformContext.Path, transformContext.Query.QueryString);
+
+            return Task.CompletedTask;
+        }
+
+        private async Task AwaitTransformRequestAsync(Task task, HttpContext httpContext, HttpRequestMessage proxyRequest, string destinationPrefix)
+        {
+            await task;
 
             if (RequestTransforms.Count == 0)
             {
@@ -95,12 +135,60 @@ namespace Microsoft.ReverseProxy.Service.RuntimeModel.Transforms
                 transformContext.DestinationPrefix, transformContext.Path, transformContext.Query.QueryString);
         }
 
-        public override async Task TransformResponseAsync(HttpContext httpContext, HttpResponseMessage proxyResponse)
+        private async Task AwaitTransformRequestAsync(ValueTask task, int i, RequestTransformContext transformContext, HttpRequestMessage proxyRequest)
+        {
+            await task;
+
+            for (i++; i < RequestTransforms.Count; i++)
+            {
+                await RequestTransforms[i].ApplyAsync(transformContext);
+            }
+
+            // Allow a transform to directly set a custom RequestUri.
+            proxyRequest.RequestUri ??= RequestUtilities.MakeDestinationAddress(
+                transformContext.DestinationPrefix, transformContext.Path, transformContext.Query.QueryString);
+        }
+
+        public override Task TransformResponseAsync(HttpContext httpContext, HttpResponseMessage proxyResponse)
         {
             if (ShouldCopyResponseHeaders.GetValueOrDefault(true))
             {
-                await base.TransformResponseAsync(httpContext, proxyResponse);
+                var task = base.TransformResponseAsync(httpContext, proxyResponse);
+                if (!task.IsCompletedSuccessfully)
+                {
+                    return AwaitTransformResponseAsync(task, httpContext, proxyResponse);
+                }
             }
+
+            if (ResponseTransforms.Count == 0)
+            {
+                return Task.CompletedTask;
+            }
+
+            var transformContext = new ResponseTransformContext()
+            {
+                HttpContext = httpContext,
+                ProxyResponse = proxyResponse,
+                HeadersCopied = ShouldCopyResponseHeaders.GetValueOrDefault(true),
+            };
+
+            for (var i = 0; i < ResponseTransforms.Count; i++)
+            {
+                var task = ResponseTransforms[i].ApplyAsync(transformContext);
+                if (!task.IsCompletedSuccessfully)
+                {
+                    return AwaitTransformResponseAsync(task, i, transformContext);
+                }
+
+                task.GetAwaiter().GetResult();
+            }
+
+            return Task.CompletedTask;
+        }
+
+        private async Task AwaitTransformResponseAsync(Task task, HttpContext httpContext, HttpResponseMessage proxyResponse)
+        {
+            await task;
 
             if (ResponseTransforms.Count == 0)
             {
@@ -120,12 +208,61 @@ namespace Microsoft.ReverseProxy.Service.RuntimeModel.Transforms
             }
         }
 
-        public override async Task TransformResponseTrailersAsync(HttpContext httpContext, HttpResponseMessage proxyResponse)
+        private async Task AwaitTransformResponseAsync(ValueTask task, int i, ResponseTransformContext transformContext)
+        {
+            await task;
+
+            for (i++; i < ResponseTransforms.Count; i++)
+            {
+                await ResponseTransforms[i].ApplyAsync(transformContext);
+            }
+        }
+
+        public override Task TransformResponseTrailersAsync(HttpContext httpContext, HttpResponseMessage proxyResponse)
         {
             if (ShouldCopyResponseTrailers.GetValueOrDefault(true))
             {
-                await base.TransformResponseTrailersAsync(httpContext, proxyResponse);
+                var task = base.TransformResponseTrailersAsync(httpContext, proxyResponse);
+                if (!task.IsCompletedSuccessfully)
+                {
+                    return AwaitTransformResponseTrailersAsync(task, httpContext, proxyResponse);
+                }
             }
+
+            if (ResponseTrailerTransforms.Count == 0)
+            {
+                return Task.CompletedTask;
+            }
+
+            // Only run the transforms if trailers are actually supported by the client response.
+            var responseTrailersFeature = httpContext.Features.Get<IHttpResponseTrailersFeature>();
+            var outgoingTrailers = responseTrailersFeature?.Trailers;
+            if (outgoingTrailers != null && !outgoingTrailers.IsReadOnly)
+            {
+                var transformContext = new ResponseTrailersTransformContext()
+                {
+                    HttpContext = httpContext,
+                    ProxyResponse = proxyResponse,
+                    HeadersCopied = ShouldCopyResponseTrailers.GetValueOrDefault(true),
+                };
+
+                for (var i = 0; i < ResponseTrailerTransforms.Count; i++)
+                {
+                    var task = ResponseTrailerTransforms[i].ApplyAsync(transformContext);
+                    if (!task.IsCompletedSuccessfully)
+                    {
+                        return AwaitTransformResponseTrailersAsync(task, i, transformContext);
+                    }
+                    task.GetAwaiter().GetResult();
+                }
+            }
+
+            return Task.CompletedTask;
+        }
+
+        private async Task AwaitTransformResponseTrailersAsync(Task task, HttpContext httpContext, HttpResponseMessage proxyResponse)
+        {
+            await task;
 
             if (ResponseTrailerTransforms.Count == 0)
             {
@@ -148,6 +285,16 @@ namespace Microsoft.ReverseProxy.Service.RuntimeModel.Transforms
                 {
                     await responseTrailerTransform.ApplyAsync(transformContext);
                 }
+            }
+        }
+
+        private async Task AwaitTransformResponseTrailersAsync(ValueTask task, int i, ResponseTrailersTransformContext transformContext)
+        {
+            await task;
+
+            for (i++; i < ResponseTrailerTransforms.Count; i++)
+            {
+                await ResponseTrailerTransforms[i].ApplyAsync(transformContext);
             }
         }
     }