Fixes https://github.com/microsoft/reverse-proxy/issues/663