@MihaZupan 2021-02-23
Did several runs, seems to be a measurable improvement of around 0.5 - 1 %.