@MihaZupan 2021-02-23
Could we reduce some code duplication by extracting the shared portion after `ShouldCopyRequestHeaders` into a separate method?