@Kahbazi 2021-02-23
> Could we reduce some code duplication by extracting the shared portion after `ShouldCopyRequestHeaders` into a separate method?

I'm not sure how to refactor it. The types are different, and each method is accessing a different property.