@Tratcher 2021-02-24
@MihaZupan did it help with allocations any?