@MihaZupan 2021-02-25
> did it help with allocations any?

No - since these methods never actually "go async", the state machines stay as structs and don't allocate.


Seeing as the performance difference is small (almost within the margin of error), I don't think it is worth duplicating all this logic.
Thank you for looking into this potential optimization @Kahbazi, but I don't think we'll be taking this change unless we see a bigger perf difference / find a way to avoid complicating the logic too much.

Did you by any chance run a microbenchmark for just the transformer to see a perf difference there?