@Kahbazi 2021-02-25
> Did you by any chance run a microbenchmark for just the transformer to see a perf difference there?

No. I will run one and post the result.