@Kahbazi 2021-02-25
Here's a benchmark of `TransformRequestAsync` with the default transformers which are removing the incoming host header and X-Forwarded headers.

|  Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| Current | 3.210 us | 0.0450 us | 0.0399 us | 0.5684 |     - |     - |   1.74 KB |
|     New | 2.894 us | 0.0267 us | 0.0208 us | 0.5531 |     - |     - |    1.7 KB |